# openapi.yaml
openapi: 3.1.0
info:
  title: Hooter Backend API
  version: '1.0.0'
  description: Internal API for Hooter backend
servers:
  - url: https://hooter_staging.h0oter.com
    description: Staging server

  - url: http://localhost:8888
    description: Local development server
...
paths:
...
  /inventory/inward/count:
    get:
      summary:  Get the inward count
      description: Get the total number of inward based upon total, completed, partial or pending
      tags:
        - Inventory
      security:
        - cookieAuth: []
      responses:
        "200":
          description:  Get the count for the inventory inward
          content:
            appliation/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 10
                  completed:
                    type: integer
                    example: 20
                  partial:
                    type: integer
                    example: 2
                  pending:
                    type: integer
                    example: 0
        "500":
          description: Inernal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
                    example: Internal server error
        default:
          $ref: "#/components/responses/default"
  /inventory/inward:
    post:
      summary: Create new Inward
      description:  Registering new inward for the product stock
      tags:
        - Inventory
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/csrf'
      requestBody:
        description:  Inward information
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                supplier_id:
                  type: integer
                  example: 10
                usku_ids:
                  type: array
                  items:
                    type: object
                    properties:
                      usku_id: 
                        type: string
                        example: 3419xxxxxxxxx
                      # shipment:
                      # warehouse_id:
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
  responses:
...
    "default":
      description: Default response
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - message
            properties:
              status:
                type: string
              message:
                type: string
  parameters:
...
    csrf:
        name: X-CSRF-Token
        in: header
        required: true
        description: CSRF acesss token acquired using /security/csrf API
        schema:
          type: string